Bugs de profile y botón de log out arreglados

+++ app/src/main/java/com/sugarcoach/ui/profile/view/ProfileHolder.kt
@@ -13,7 +13,7 @@ import java.util.*
 class ProfileHolder(private val binding: ProfileItemBinding) : RecyclerView.ViewHolder(binding.root) {
 
     fun bind(item: ProfileItem, activity: ProfileActivity) {
-        binding.profileAvatar.setImageDrawable(activity.getDrawable(item.image.toInt()))
+        binding.profileAvatar.setImageDrawable(activity.getDrawable(item.image))
         binding.profileAvatar.borderColor = ContextCompat.getColor(activity, R.color.transparent)
         if (item.selected) {
             binding.profileAvatar.borderColor = ContextCompat.getColor(activity, R.color.purple)
